允许用户查看和进行评价

+++ cloud-demo/shop-service/src/main/java/cn/itcast/order/repository/PurchaseRecordRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PurchaseRecordRepository extends JpaRepository<PurchaseRecord, Long> {
+public interface PurchaseRecordRepository extends JpaRepository<PurchaseRecord, Integer> {
     List<PurchaseRecord> findByBuyerId(String buyerId);
     List<PurchaseRecord> findByGoodsId(String goodsId);
 }
